#S3 Fix panic and ride route update

+++ backend/src/main/java/com/tricolori/backend/dto/osrm/OSRMRouteResponse.java
@@ -12,7 +12,7 @@ public class OSRMRouteResponse {
     @Data
     public static class OSRMRoute {
         private Double distance; // in meters
-        private Double duration; // in seconds
+        private Long duration; // in seconds
         private String geometry; // encoded polyline
     }
 }
